requirements.txt
# ============================
# Core API & Server
# ============================
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-multipart==0.0.9

# ============================
# Core ML & Data
# ============================
numpy==1.26.4
pandas==2.2.1
joblib==1.3.2
scikit-learn==1.6.1

# ============================
# Deep Learning (TensorFlow stack)
# ============================
tensorflow==2.16.1
keras==3.12.0
tf-keras==2.16.0

# ============================
# Torch (CPU only)
# ============================
torch>=2.0,<3.0

# ============================
# Classical ML
# ============================
xgboost==2.0.3

# ============================
# RAG / GenAI
# ============================
faiss-cpu==1.13.1
sentence-transformers==5.2.0
transformers==4.57.3
huggingface-hub==0.36.0
pypdf==6.4.2
openai==2.13.0
codecarbon==3.2.0

# ============================
# Visualization / Frontend
# ============================
matplotlib==3.8.3
seaborn==0.13.2
streamlit==1.39.0

# ============================
# Utilities
# ============================
tqdm==4.66.2
pyyaml==6.0.1
python-dotenv==1.2.1
requests==2.32.5

# ============================
# PDF / Reporting
# ============================
reportlab==4.1.0
